# src/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API MemoBelc",
    "description": "Documentação da API MemoBelc. Use **Authorize** para informar o token Bearer e testar rotas protegidas."
  },
  "host": "127.0.0.1:5000",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Token JWT: use o valor 'Bearer <seu_token>'"
    }
  },
  "security": [],
...
  "paths": {
...
    "/billing/me": {
      "get": {
        "tags": ["Billing"],
        "summary": "Assinatura e entitlements do usuário",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Estado da assinatura" } }
      }
    },
    "/billing/payments": {
      "get": {
        "tags": ["Billing"],
        "summary": "Histórico de pagamentos do usuário",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Pagamentos" } }
      }
    },
...
    "/admin/billing/subscriptions": {
      "get": {
        "tags": ["AdminBilling"],
        "summary": "Listar assinaturas com filtros",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Assinaturas" } }
      }
    },
    "/admin/billing/payments": {
      "get": {
        "tags": ["AdminBilling"],
        "summary": "Listar compras e transações",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Pagamentos" } }
      }
    },
    "/admin/billing/grants": {
      "post": {
        "tags": ["AdminBilling"],
        "summary": "Liberar plano, livro, conjunto ou serviço manualmente",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Concedido" } }
      }
    },
    "/admin/billing/external-sales": {
      "get": {
        "tags": ["AdminBilling"],
        "summary": "Listar vendas externas",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Vendas" } }
      },
      "post": {
        "tags": ["AdminBilling"],
        "summary": "Registrar venda externa e conceder acesso",
        "security": [{ "Bearer": [] }],
        "responses": { "201": { "description": "Registrada" } }
      }
    },
    "/admin/billing/access": {
      "get": {
        "tags": ["AdminBilling"],
        "summary": "Regras de visibilidade dos serviços",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Regras" } }
      }
    },
    "/admin/billing/classrooms": {
      "get": {
        "tags": ["AdminBilling"],
        "summary": "Listar turmas e flags de checkout",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Turmas" } }
      }
    },
    "/admin/billing/classrooms/{classroom_id}": {
      "put": {
        "tags": ["AdminBilling"],
        "summary": "Permitir ou bloquear checkout da turma",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "classroom_id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "checkout_allowed": { "type": "boolean" }
              }
            }
          }
        ],
        "responses": { "200": { "description": "Turma atualizada" } }
      }
    },
    "/admin/billing/classroom-checkouts": {
      "get": {
        "tags": ["AdminBilling"],
        "summary": "Histórico e comprovantes de checkout de turmas",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Pagamentos com comprador, turma e receipt_url" } }
      }
    }
  }
}